LPS-96442 Reduce FlatJSBundle heap usage by removing most of the fields from JSBundleAdapter.
